Move rcbridge initialization to RcloneProvider
RSAF was previously initializing rcbridge in MainApplication.onCreate().
Since all ContentProviders are initialized before the Application, it's
possible to start receiving ContentProvider requests before rcbridge is
initialized. This seems to trigger a deadlock in rclone and causes
Android to kill RSAF due to the main thread hanging. When hitting this
issue, crash.log would not be written because the killing is done via
SIGQUIT, not via a JVM exception. This commit fixes the issue by
initializing rcbridge in RcloneProvider.onCreate().

Fixes: #22
